# tests/contract/api/openapi_snapshot.json
{
 "components": {
  "schemas": {
...
   "ClaimEvidenceView": {
    "additionalProperties": false,
    "description": "\u4e00\u6761 claim \u7684\u6295\u5f71\uff1aclaim_ref \u5fc5\u6709\uff1b\u5176\u4f59\u5b57\u6bb5\u53ea\u5728 canonical \u8f7d\u8377\u643a\u5e26\u65f6\u51fa\u73b0\u3002",
    "properties": {
     "canonical_value": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Canonical Value"
     },
     "claim_ref": {
      "title": "Claim Ref",
      "type": "string"
     },
     "claim_type": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Claim Type"
     },
     "evidence_refs": {
      "items": {
       "additionalProperties": true,
       "type": "object"
      },
      "title": "Evidence Refs",
      "type": "array"
     },
     "quote_text": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Quote Text"
     },
     "verified": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Verified"
     }
    },
    "required": [
     "claim_ref"
    ],
    "title": "ClaimEvidenceView",
    "type": "object"
   },
   "ClaimScope": {
    "additionalProperties": false,
    "description": "claim \u7684\u53e3\u5f84\u8fb9\u754c\uff1amode/model/version/date/corpus/environment \u5168\u90e8\u8fdb\u6bd4\u5bf9\u3002",
    "properties": {
     "corpus": {
      "minLength": 1,
      "title": "Corpus",
      "type": "string"
     },
     "date": {
      "minLength": 1,
      "title": "Date",
      "type": "string"
     },
     "environment": {
      "minLength": 1,
      "title": "Environment",
      "type": "string"
     },
     "mode": {
      "minLength": 1,
      "title": "Mode",
      "type": "string"
     },
     "model": {
      "minLength": 1,
      "title": "Model",
      "type": "string"
     },
     "version": {
      "minLength": 1,
      "title": "Version",
      "type": "string"
     }
    },
    "required": [
     "mode",
     "model",
     "version",
     "date",
     "corpus",
     "environment"
    ],
    "title": "ClaimScope",
    "type": "object"
   },
   "ClaimView": {
    "additionalProperties": false,
    "properties": {
     "bound_value": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Bound Value"
     },
     "claim_id": {
      "title": "Claim Id",
      "type": "string"
     },
     "evidence": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Evidence"
     },
     "scope": {
      "additionalProperties": true,
      "title": "Scope",
      "type": "object"
     },
     "statement": {
      "title": "Statement",
      "type": "string"
     },
     "status": {
      "title": "Status",
      "type": "string"
     }
    },
    "required": [
     "claim_id",
     "statement",
     "scope",
     "status",
     "bound_value",
     "evidence"
    ],
    "title": "ClaimView",
    "type": "object"
   },
...
   "ConflictEvidenceView": {
    "additionalProperties": false,
    "description": "ADR-005 \u51b2\u7a81\u8bb0\u5f55\u6295\u5f71\uff08\u5e76\u5217\u4fdd\u7559\u53cc\u65b9\u53d6\u503c\uff0c\u4e0d\u4ef2\u88c1\uff09\u3002",
    "properties": {
     "detected_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Detected At"
     },
     "evidence_refs": {
      "items": {
       "type": "string"
      },
      "title": "Evidence Refs",
      "type": "array"
     },
     "field": {
      "title": "Field",
      "type": "string"
     },
     "values": {
      "additionalProperties": true,
      "title": "Values",
      "type": "object"
     }
    },
    "required": [
     "field",
     "values"
    ],
    "title": "ConflictEvidenceView",
    "type": "object"
   },
...
   "CreateRunRequest": {
    "additionalProperties": false,
    "description": "POST /runs \u7684\u8bf7\u6c42\u4f53\uff08planner \u610f\u56fe\uff09\u3002",
    "properties": {
     "template": {
      "default": "single-fixture",
      "maxLength": 64,
      "title": "Template",
      "type": "string"
     },
     "workspace_id": {
      "format": "uuid",
      "title": "Workspace Id",
      "type": "string"
     }
    },
    "required": [
     "workspace_id"
    ],
    "title": "CreateRunRequest",
    "type": "object"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "RegisterClaimRequest": {
    "additionalProperties": false,
    "properties": {
     "claim_id": {
      "title": "Claim Id",
      "type": "string"
     },
     "scope": {
      "$ref": "#/components/schemas/ClaimScope"
     },
     "statement": {
      "title": "Statement",
      "type": "string"
     }
    },
    "required": [
     "claim_id",
     "statement",
     "scope"
    ],
    "title": "RegisterClaimRequest",
    "type": "object"
   },
...
   "RunEvidenceView": {
    "additionalProperties": false,
    "description": "\u4e00\u4e2a run \u7684 evidence/claim \u6295\u5f71\uff08canonical \u4e8b\u4ef6\u771f\u76f8 + ADR-006 \u53ef\u89c1\u6027\uff09\u3002",
    "properties": {
     "answer": {
      "additionalProperties": true,
      "title": "Answer",
      "type": "object"
     },
     "answer_status": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Answer Status"
     },
     "claims": {
      "items": {
       "$ref": "#/components/schemas/ClaimEvidenceView"
      },
      "title": "Claims",
      "type": "array"
     },
     "clarification": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Clarification"
     },
     "conflicts": {
      "items": {
       "$ref": "#/components/schemas/ConflictEvidenceView"
      },
      "title": "Conflicts",
      "type": "array"
     },
     "failed_claims": {
      "items": {
       "type": "string"
      },
      "title": "Failed Claims",
      "type": "array"
     },
     "findings": {
      "items": {},
      "title": "Findings",
      "type": "array"
     },
     "run_id": {
      "format": "uuid",
      "title": "Run Id",
      "type": "string"
     },
     "run_status": {
      "title": "Run Status",
      "type": "string"
     },
     "unknowns": {
      "items": {
       "type": "string"
      },
      "title": "Unknowns",
      "type": "array"
     },
     "verification": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Verification"
     },
     "verified_claims": {
      "items": {
       "type": "string"
      },
      "title": "Verified Claims",
      "type": "array"
     }
    },
    "required": [
     "run_id",
     "run_status"
    ],
    "title": "RunEvidenceView",
    "type": "object"
   },
   "RunRecord": {
    "additionalProperties": false,
    "description": "\u4e00\u4e2a run \u7684\u5217\u8868\u9879\u6295\u5f71\uff08\u6765\u81ea PG reduce\uff09\u3002",
    "properties": {
     "organization_id": {
      "format": "uuid",
      "title": "Organization Id",
      "type": "string"
     },
     "run_id": {
      "format": "uuid",
      "title": "Run Id",
      "type": "string"
     },
     "status": {
      "title": "Status",
      "type": "string"
     },
     "template": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Template"
     }
    },
    "required": [
     "run_id",
     "status",
     "organization_id"
    ],
    "title": "RunRecord",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
...
  }
 },
 "info": {
  "title": "zhiwei",
  "version": "0.1.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/v1/claims": {
   "get": {
    "operationId": "list_claims_api_v1_claims_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/ClaimView"
         },
         "title": "Response List Claims Api V1 Claims Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "summary": "List Claims",
    "tags": [
     "claims"
    ]
   },
   "post": {
    "operationId": "register_claim_api_v1_claims_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RegisterClaimRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ClaimView"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Register Claim",
    "tags": [
     "claims"
    ]
   }
  },
...
  "/api/v1/runs": {
   "get": {
    "operationId": "list_runs_api_v1_runs_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/RunRecord"
         },
         "title": "Response List Runs Api V1 Runs Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "summary": "List Runs",
    "tags": [
     "runs"
    ]
   },
   "post": {
    "operationId": "create_run_api_v1_runs_post",
    "parameters": [
     {
      "in": "header",
      "name": "Idempotency-Key",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "minLength": 1,
         "pattern": "\\S+",
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Idempotency-Key"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CreateRunRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Idempotent replay (Idempotency-Key \u91cd\u8bd5\u547d\u4e2d\u5df2\u521b\u5efa Run)"
     },
     "201": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Create Run",
    "tags": [
     "runs"
    ]
   }
  },
...
  "/api/v1/runs/{run_id}/evidence": {
   "get": {
    "operationId": "get_run_evidence_api_v1_runs__run_id__evidence_get",
    "parameters": [
     {
      "in": "path",
      "name": "run_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Run Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RunEvidenceView"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Run Evidence",
    "tags": [
     "evidence"
    ]
   }
  },
...
 }
}